refactor(core): 将 ThreadHelper 迁出 helper 至 app.core.thread，消除 core→helper 反向依赖
ThreadHelper 是依赖 settings 的线程池基础设施，却位于 app/helper/thread.py，导致
core/event.py 顶层反向依赖 helper 层（架构审计 P1：core 应为零 helper 依赖的稳定底座）。
注意 ThreadHelper 依赖 settings，故不可下沉到 utils（会形成 utils→core 反向边），其正确
归属是 core 自身。

改动：
- 新增 app/core/thread.py：ThreadHelper 的权威新家（内容不变；仅依赖 core.config 与
  utils.singleton，方向正确）。
- app/helper/thread.py 改为 re-export 兼容垫片，保留历史导入路径（无插件依赖该模块，
  纯防御性兼容）。
- 5 处内部 importer 改为 from app.core.thread import ThreadHelper：
  core/event、command、chain/download、modules/telegram、startup/modules_initializer。
- 新增 test_core_thread_relocation.py：垫片同类 / submit 可用 / core/event 不再引用
  helper.thread / core.thread 无 helper 依赖。

效果：消除 core/event→helper.thread top-level 反向边；ThreadHelper 行为零变化。
（event.py:706 的 lazy helper.message 属另一接缝，且 MessageHelper 是插件注入 API，本 PR 不动。）
