+++ app/command.py
@@ -13,8 +13,8 @@ from app.chain.system import SystemChain
 from app.chain.transfer import TransferChain
 from app.core.event import Event as ManagerEvent, eventmanager, Event
 from app.core.plugin import PluginManager
+from app.core.thread import ThreadHelper
 from app.helper.message import MessageHelper
-from app.helper.thread import ThreadHelper
 from app.log import logger
 from app.scheduler import Scheduler
 from app.schemas import Notification, CommandRegisterEventData
