+++ app/helper/thread.py
@@ -1,29 +1,5 @@
-from concurrent.futures import ThreadPoolExecutor
+# 兼容垫片：ThreadHelper 已迁移至 app.core.thread（线程池属核心基础设施且依赖 settings，
+# 不应位于 helper 层而被 core 反向依赖）。此处保留 re-export 以兼容历史导入路径。
+from app.core.thread import ThreadHelper
 
-from app.core.config import settings
-from app.utils.singleton import Singleton
-
-
-class ThreadHelper(metaclass=Singleton):
-    """
-    线程池管理
-    """
-    def __init__(self):
-        self.pool = ThreadPoolExecutor(max_workers=settings.CONF.threadpool)
-
-    def submit(self, func, *args, **kwargs):
-        """
-        提交任务
-        :param func: 函数
-        :param args: 参数
-        :param kwargs: 参数
-        :return: future
-        """
-        return self.pool.submit(func, *args, **kwargs)
-
-    def shutdown(self):
-        """
-        关闭线程池
-        :return:
-        """
-        self.pool.shutdown()
+__all__ = ["ThreadHelper"]
